Fixed bug when updating questions in question bank

+++ core/views/code_questions.py
@@ -161,8 +161,10 @@ def update_test_cases(request, code_question_id):
     if request.method == "POST":
         testcase_formset = TestCaseFormset(request.POST, instance=code_question, prefix='tc')
         if testcase_formset.is_valid():
+
             # remove past attempts
-            assessment_attempts = code_question.assessment.assessmentattempt_set.all().delete()
+            if code_question.assessment:
+                code_question.assessment.assessmentattempt_set.all().delete()
 
             testcase_formset.save()
             messages.success(request, "Test cases updated!")
@@ -203,8 +205,10 @@ def update_languages(request, code_question_id):
     if request.method == "POST":
         code_snippet_formset = CodeSnippetFormset(request.POST, instance=code_question, prefix='cs')
         if code_snippet_formset.is_valid():
+
             # remove past attempts
-            assessment_attempts = code_question.assessment.assessmentattempt_set.all().delete()
+            if code_question.assessment:
+                code_question.assessment.assessmentattempt_set.all().delete()
 
             code_snippet_formset.save()
             messages.success(request, "Code Snippets saved!")
